Fix a problem with getStorageRoot() yielding the real_path
There was a problem if the media-dir was behind a symlink that caused
wysiwyg images to be synched multiple times becuase the path from the
cloudinary_synchronisation table was always diffrent from the one that
was in the filesystem-collection. Because of this every image was
synched every time the sync tried to run.