SESSCLOUD-419 Remove folder reference during delete operation
This change prevents the image from having a relative path when in non-foldered mode.
This prevents an unwanted forward slash when the SDK is called that causes the
delete operation to fail.
This could be implemented by refactoring but that will take considerably more effort
to ensure no current features are effected.